requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
whitenoise
django-cors-headers
sendgrid-django  # optional: use when you want production email via SendGrid
